+++ src/oersted/magnetization.py
@@ -4,7 +4,7 @@ from collections.abc import Callable
 
 import numpy as np
 from numpy.typing import NDArray
-from numpy import float64, int64
+from numpy import float64, uint32
 
 from oersted import LinearMaterial, MU0, Mesh
 from oersted.materials import Material
@@ -65,7 +65,7 @@ def mag_force(centroids: NDArray[float64], vol: NDArray[float64], material: Line
 
 
 def h_demag_tet4(
-    nodes: NDArray[float64], element_connectivity: NDArray[int64], material: Material, m_field: NDArray[float64], nthreads_requested: int = 0
+    nodes: NDArray[float64], element_connectivity: NDArray[uint32], material: Material, m_field: NDArray[float64], nthreads_requested: int = 0
 ) -> NDArray[float64]:
     """Compute the demagnetization field H(M) on mesh element centroids given the current M-field
 
@@ -110,7 +110,7 @@ def h_demag_tet4(
 
 def h_demag_tet4_octree(
     nodes: NDArray[float64],
-    element_connectivity: NDArray[int64],
+    element_connectivity: NDArray[uint32],
     material: Material,
     m_field: NDArray[float64],
     centroids: NDArray[float64],