+++ CHANGELOG.md
@@ -37,3 +37,4 @@
 * Added Mesh object, methods for calculating volume and centroid of tet4 elements
 * Update magnetization functions and related tests to use Mesh object
 * Reduce default error tolerance of demag calc to 1.0 without loss of Bfield precision to 6 decimal places
+* Add function for computing which are surface faces on a tetrahedral mesh